# schema/openapi/openapi.yaml
openapi: '3.0.3'
info:
  title: Backbone API
  version: '1.0.0'
  description: |
    Auto-generated by backbone-cli.
    This API provides 11 standard CRUD endpoints for each model.
servers:
  - url: http://localhost:3000
    description: Development server
...
paths:
...
  /api/v1/stock_reconciliation_items:
    get:
      tags:
        - StockReconciliationItems
      summary: List StockReconciliationItems
      description: Get a paginated list of StockReconciliationItems with optional filtering and sorting
      operationId: listStockReconciliationItems
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockReconciliationItemList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - StockReconciliationItems
      summary: Create StockReconciliationItem
      description: Create a new StockReconciliationItem
      operationId: createStockReconciliationItem
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateStockReconciliationItemRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StockReconciliationItem'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
  /api/v1/warehouses:
    get:
      tags:
        - Warehouses
      summary: List Warehouses
      description: Get a paginated list of Warehouses with optional filtering and sorting
      operationId: listWarehouses
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/PerPageParam'
        - $ref: '#/components/parameters/SortByParam'
        - $ref: '#/components/parameters/SortOrderParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WarehouseList'
        '401':
          $ref: '#/components/responses/Unauthorized'

    post:
      tags:
        - Warehouses
      summary: Create Warehouse
      description: Create a new Warehouse
      operationId: createWarehouse
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateWarehouseRequest'
      responses:
        '201':
          description: Created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Warehouse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  schemas:
...
    StockReconciliationItem:
      type: object
      required:
        - id
        - reconciliation_id
        - company_id
        - item_id
        - counted_qty
        - counted_rate
        - qty_difference
        - value_difference
        - metadata
      properties:
        id:
          type: string
          format: uuid
        reconciliation_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        counted_qty:
          type: string
        counted_rate:
          type: string
        qty_difference:
          type: string
        value_difference:
          type: string
        metadata:
          type: object
...
    CreateStockReconciliationItemRequest:
      type: object
      required:
        - reconciliation_id
        - company_id
        - item_id
        - counted_qty
        - counted_rate
        - qty_difference
        - value_difference
        - metadata
      properties:
        reconciliation_id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        item_id:
          type: string
          format: uuid
        counted_qty:
          type: string
        counted_rate:
          type: string
        qty_difference:
          type: string
        value_difference:
          type: string
        metadata:
          type: object
...
    StockReconciliationItemList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/StockReconciliationItem'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
    Warehouse:
      type: object
      required:
        - id
        - company_id
        - code
        - name
        - warehouse_type
        - is_group
        - metadata
      properties:
        id:
          type: string
          format: uuid
        company_id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 40
        name:
          type: string
          maxLength: 150
        warehouse_type:
          type: string
        parent_warehouse_id:
          nullable: true
          type: string
          format: uuid
        is_group:
          type: boolean
        metadata:
          type: object
...
    CreateWarehouseRequest:
      type: object
      required:
        - company_id
        - code
        - name
        - warehouse_type
        - is_group
        - metadata
      properties:
        company_id:
          type: string
          format: uuid
        code:
          type: string
          maxLength: 40
        name:
          type: string
          maxLength: 150
        warehouse_type:
          type: string
        parent_warehouse_id:
          nullable: true
          type: string
          format: uuid
        is_group:
          type: boolean
        metadata:
          type: object
...
    WarehouseList:
      type: object
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/Warehouse'
        meta:
          $ref: '#/components/schemas/PaginationMeta'
...
    PaginationMeta:
      type: object
      properties:
        total:
          type: integer
          example: 100
        page:
          type: integer
          example: 1
        per_page:
          type: integer
          example: 20
        total_pages:
          type: integer
          example: 5
    Error:
      type: object
      required:
        - code
        - message
      properties:
        code:
          type: string
          example: VALIDATION_ERROR
        message:
          type: string
          example: Invalid input data
        details:
          type: object
          additionalProperties: true
  parameters:
...
    PageParam:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
    PerPageParam:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
    SortByParam:
      name: sort_by
      in: query
      schema:
        type: string
    SortOrderParam:
      name: sort_order
      in: query
      schema:
        type: string
        enum: [asc, desc]
        default: asc
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - bearerAuth: []